+++ src/scripts/crawl.cjs
@@ -11,37 +11,52 @@ const { resolve, sep } = require("path");
  * @returns {string[]}
  */
 function crawl(directory, filters = "both", maxDepth = Infinity) {
+  /** @type {string[]} */
+  const files = [];
   const directories = [directory];
   const temp = [directory];
 
   for (let i = 0; i < maxDepth; i++) {
-    const subDirectories = temp
+    const directoryContents = temp
       .map((path) =>
         readdirSync(resolve(__dirname, path), { withFileTypes: true })
       )
-      .flat(1)
-      .filter(
-        (directory) =>
-          filters === "both" ||
-          (filters === "directories" && directory.isDirectory()) ||
-          (filters === "files" && !directory.isDirectory())
-      )
-      .map((directory) => `${directory.path}${sep}${directory.name}`);
+      .flat(1);
+
+    const directoryFiles =
+      filters === "files" || filters === "both"
+        ? directoryContents.filter((directory) => directory.isFile())
+        : [];
+    const subDirectories = directoryContents.filter((directory) =>
+      directory.isDirectory()
+    );
+
+    const subDirectoryNames = subDirectories.map(
+      (directory) => `${directory.path}${sep}${directory.name}`
+    );
+    const fileNames = directoryFiles.map(
+      (directory) => `${directory.path}${sep}${directory.name}`
+    );
 
     if (i === 0) {
       temp.splice(0, 1);
     }
 
-    directories.push(...subDirectories);
+    directories.push(...subDirectoryNames);
+    files.push(...fileNames);
     temp.splice(0, temp.length);
-    temp.push(...subDirectories);
+    temp.push(...subDirectoryNames);
 
-    if (subDirectories.length === 0) {
+    if (directoryContents.length === 0) {
       break;
     }
   }
 
-  return directories;
+  return filters === "both"
+    ? [...directories, ...files]
+    : filters === "directories"
+      ? directories
+      : files;
 }
 
 module.exports = crawl;